requirements.txt
streamlit==1.29.0
langchain==0.1.4
langchain-google-genai==0.0.8
google-generativeai==0.3.2
opencv-python==4.8.1.78
pillow==10.1.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.4